Change css styling for navbar links
Change from width being 200px to the padding of each link being 200px.
The reason is because padding makes sense if user wants to click on the section of link
rather than the text.
So I created new custom class for that since it can be used if more links are to be added
in the navbar.