Separate login and register errors from one default error
This was done so that each field could have errors for them separately.
For example, if the user entered an invalid Email like "email@12" and submitted it, the "Invalid Email format" or
something like that could show up only for Email field and not for other fields.

Also makes the code easier to maintain and debug by separating them.