Remove the extra password validation from login form
Only check if the password is empty or not in the login form because it
doesn't make sense checking if the password has more than six letters, or
if it has at least one number and letter, etc in login form.

That belongs in registration form while creating data, not while checking it.