Reset general error field on each submit
This is done so that error is focused on only one thing, either validation or network.
We do not want network error to constantly appear if the user has somehow got the network back but failed the validation.

TLDR; either only validation error will be displayed or network error, not both at the same time