Create a handleFormSubmit function and abstract the API calling code inside that function
The reason to create this function is so that it can be dynamic for all kinds of forms.
Mainly to abstract and encapsulate the code and to follow good pracitce (DRY).
This function is only supposed to handle the API calling part for forms, for example, sending form data or receiving the data.