Make the show/hide password button explicitly as button type
This fixes the glitch when you submit the form by pressing Enter key.
Instead of submitting the form, it would toggle the show password on and off on Enter press.

That is not what we want, we want to submit the form on Enter press, which is what happens now.