Only send the JWT string instead of sending email, username and id directly
JWT will already contain those information so I will decode them in the frontend and use them.
It makes the authorisation process more simple and secure, in my opinion.

Because we are just using one single source instead of multiple (email, username, etc), it makes the debugging easier as well.