Handle the error's response data being empty
The response.data usually does have some data but, if the backend code crashes because of some logical error.
It being empty means the error field for general error is also empty so users have no way of telling if something went wrong.

Now it has been handled and if the response.data is empty or null, we will display a default message.