Create FormFieldWrapper component that does the following
- Wraps the input elements. The reasons are:
First, to separate label component from input component.
Input component should technically only have input element in it. Label is a different component and can be used seprately.
Second, to use the same single Label component for different types of form elements like input and textarea.
For fields that require textarea, the same Label component can be used that is wrapped by FormFieldWrapper.

- Separate Label component to a different file.
- Create a textarea component. Just a filler component yet.

This commit is basically about create FormFieldWrapper component with the goal to follow DRY principle and reusability of Label component.